EnergyModelsLanguageInterfaces

EnergyModelsLanguageInterfaces — Module

The module EnergyModelsLanguageInterfaces is a package that provides utility functions for the EnergyModelsX framework.

This Julia package implements additional utilities that can be used in combination with EnergyModelsX.

These utilities do not directly incorporate new technology descriptions. They can be instead added to your technology type and included in a create_node method for your new types. You can also use the utility functions in both pre-processing and model building.

Manual outline

Quick Start

Examples

Release notes

Description of the new types

Utility functions

Default utility function

How to guides

Contribute to EnergyModelsLanguageInterfaces

Utilize the concepts of EnergyModelsLanguageInterfaces

Library outline

Public interface

Types

Methods - Internal

Methods - EnergyModelsBase