Methods - Internal

Index

EnergyModelsLanguageInterfaces.bio_type

EnergyModelsLanguageInterfaces.cleanup_libraries

EnergyModelsLanguageInterfaces.electricity_resource

EnergyModelsLanguageInterfaces.get_python_function

EnergyModelsLanguageInterfaces.getfirst

EnergyModelsLanguageInterfaces.moisture

EnergyModelsLanguageInterfaces.@dlsym

Utility methods

EnergyModelsLanguageInterfaces.cleanup_libraries — Function

cleanup_libraries()

Close all the C module libraries that have been loaded by EnergyModelsLanguageInterfaces.

EnergyModelsLanguageInterfaces.getfirst — Function

getfirst(f::Function, a::Vector)

Return the first element of Vector a satisfying the requirement of Function f.

EnergyModelsLanguageInterfaces.get_python_function — Function

get_python_function(module_name::String, function_name::String)

Import the requested function function_name from the python module module_name.

EnergyModelsLanguageInterfaces.moisture — Function

bio_type(p::ResourceBio)

Returns the moisture content of a ResourceBio p.

EnergyModelsLanguageInterfaces.bio_type — Function

bio_type(p::ResourceBio)

Returns the biomass type of a ResourceBio p.

EnergyModelsLanguageInterfaces.electricity_resource — Function

electricity_resource(n::BioCHP)

Returns the electricity resource of BioCHP node n.

Macros

EnergyModelsLanguageInterfaces.@dlsym — Macro

@dlsym(lib, func)

This macro uses dlsym to load a function from a shared library specified by lib and func. It caches the result in a Ref to avoid repeated lookups. If the symbol is already loaded, it returns the cached pointer. Otherwise, it loads the symbol, caches it, and then returns the pointer.